fix(business): finish code review requests

+++ src/modules/business/services/businessComment/DeleteBusinessCommentService.ts
@@ -6,6 +6,7 @@ import IBusinessCommentRepository from '../../repositories/IBusinessCommentRepos
 
 interface IRequest {
   business_comment_id: string;
+  user_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteBusinessCommentService {
     private businessCommentRepository: IBusinessCommentRepository,
   ) {}
 
-  public async execute({ business_comment_id }: IRequest): Promise<void> {
+  public async execute({
+    business_comment_id,
+    user_id,
+  }: IRequest): Promise<void> {
     const businessComment = await this.businessCommentRepository.findById(
       business_comment_id,
     );
 
     if (!businessComment) {
       throw new AppError('Comment not found.', 404);
+    }
+
+    if (businessComment.user_id !== user_id) {
+      throw new AppError(
+        'User does not have permission to delete this comment.',
+        403,
+      );
     }
 
     await this.businessCommentRepository.delete(business_comment_id);
