+++ src/modules/business/services/businessRating/DeleteBusinessRatingService.ts
@@ -6,6 +6,7 @@ import IBusinessRatingRepository from '../../repositories/IBusinessRatingReposit
 
 interface IRequest {
   business_rating_id: string;
+  user_id: string;
 }
 
 @injectable()
@@ -15,13 +16,23 @@ export default class DeleteBusinessRatingService {
     private businessRatingRepository: IBusinessRatingRepository,
   ) {}
 
-  public async execute({ business_rating_id }: IRequest): Promise<void> {
+  public async execute({
+    user_id,
+    business_rating_id,
+  }: IRequest): Promise<void> {
     const businessRating = await this.businessRatingRepository.findById(
       business_rating_id,
     );
 
     if (!businessRating) {
       throw new AppError('Rating not found.', 404);
+    }
+
+    if (businessRating.user_id !== user_id) {
+      throw new AppError(
+        'User does not have permission to delete this rating.',
+        403,
+      );
     }
 
     await this.businessRatingRepository.delete(business_rating_id);
